refactor(judicial-system): reorganize formatters into subdirectories
Split the flat formatters folder into pdfHelpers/, generatedPdfs/,
helpers/, and confirmation/ to separate concerns — string formatters,
PDF primitives, generated PDF documents, and confirmation utilities.
Also exports formatCustodyRestrictions from the public index.
