+++ apps/judicial-system/backend/src/app/formatters/confirmation/getCaseFileHash.ts
