+++ apps/judicial-system/backend/src/app/formatters/confirmation/index.ts
@@ -1,0 +1,2 @@
+export * from './confirmedPdf'
+export * from './getCaseFileHash'
