+++ apps/judicial-system/backend/src/app/formatters/generatedPdfs/indictmentCourtRecordPdf.spec.ts
