+++ apps/judicial-system/backend/src/app/formatters/generatedPdfs/verdictServiceCertificatePdf.ts
@@ -13,8 +13,8 @@ import {
   VerdictServiceStatus,
 } from '@island.is/judicial-system/types'
 
-import { serviceCertificate as strings } from '../messages'
-import { Case, Defendant, Verdict } from '../modules/repository'
+import { serviceCertificate as strings } from '../../messages'
+import { Case, Defendant, Verdict } from '../../modules/repository'
 import {
   addEmptyLines,
   addLargeHeading,
@@ -22,7 +22,7 @@ import {
   addNormalCenteredText,
   addNormalText,
   setTitle,
-} from './pdfHelpers'
+} from '../pdfHelpers'
 
 const getRole = (userRole?: UserRole) => {
   switch (userRole) {
