+++ apps/judicial-system/backend/src/app/formatters/helpers/PDFKitCoatOfArms.ts
