+++ apps/judicial-system/backend/src/app/formatters/helpers/coatOfArms.ts
