+++ apps/judicial-system/backend/src/app/formatters/helpers/policeStar.ts
