+++ apps/judicial-system/backend/src/app/formatters/pdfHelpers/pdf.ts
