+++ apps/judicial-system/backend/src/app/formatters/pdfHelpers/pdfHelpers.spec.ts
